refactor(rebalancer): change the type of multiplier from uint128 to uint256
It's to account for the number of decimals, which would cap the highest value to ~3.4x with a
uint128

+++ src/interfaces/Rebalancer/IRebalancerTypes.sol
@@ -24,8 +24,8 @@ interface IRebalancerTypes {
      */
     struct PositionData {
         uint128 amount;
-        uint128 entryAccMultiplier;
-        uint128 pnlMultiplier;
+        uint256 entryAccMultiplier;
+        uint256 pnlMultiplier;
         PositionId id;
     }
 }
